docker/python/requirements.txt
numpy==1.26.4
pandas
scikit-learn
matplotlib
flask
twilio
python-dotenv
openai
pyngrok
google-generativeai
mysql-connector-python
sqlalchemy
nltk
spacy
cython